접속 끊길 때 처리 + 서버 기타 처리 안정화(+디버깅용 클라 수 조정) + 클라 회전보간
1. 접속 끊길 때 정상적으로 서버에서 처리되고, 클라에서는 정상적으로 오브젝트는 삭제되지만 그 오브젝트를 참조하고있는 다른 오브젝트들에 대한 처리가 필요함.

2. 기타 Room 부분 대폭 수정, 섹터에 담길 STL List 추가.
C_Global.cpp 파일에 MAX_PLAYER의 숫자를 조정하면 원하는 플레이어 수 만큼으로 디버깅 할 수 있음!!

3. 클라이언트의 회전정보를 PositionPacket에 담아 보내며, 클라측에서는 Quarternion.Lerp() 함수를 이용해서 회전 보간을 함.(패킷 송신 주기는 0.14초)

4. 클라에서 송신하는 포지션을 World -> Local로 변환했고, 이제 이를 서버에서 섹터에서 처리하도록 해야함.

5. 방 여러개 생성되고, 소멸되는 과정처리 됨

<추가해야할 것>
1.  플레이어가 나갔을 때 에디터에서 MissingReferenceException이 발생하는데 이에 대한 예외 처리

2. 정상종료의 경우에는 종료 잘되고, 서버에서 종료 처리도 잘 하지만 와이파이가 끊겼거나 디바이스가 끊겨버린 경우에는 짧게는 1분에서 길게는 언제까지 처리될지 모르는 상황이 오는데 이거를 AliveCheck같은 타이머 쓰레드를 놓고 주기적으로 패킷을 보내야할지 고민해봐야 함.

3. 클라에서 반복되는 부분에 new를 통해 생성되는 가비지를 최적화 해야할 것 같음

4. 클라 프로퍼티 부분은 get; set; 축약형으로 쓰는게 가독성이 더 좋을 것 같음

+++ Server/Server/UtilityManager.cpp
@@ -43,8 +43,8 @@ void UtilityManager::UnicodeToUTF8(wchar_t* _strUnicode, char* _strUTF8)
 void UtilityManager::UTF8ToUnicode(char* _strUTF8, wchar_t* _strUnicode)
 {
 	//_strUnicode = CA2W(_strUTF8, CP_UTF8);
-	int nLen = MultiByteToWideChar(CP_UTF8, 0, _strUTF8, strlen(_strUTF8), NULL, NULL);
-	MultiByteToWideChar(CP_UTF8, 0, _strUTF8, strlen(_strUTF8), _strUnicode, nLen);
+	int nLen = MultiByteToWideChar(CP_UTF8, 0, _strUTF8, (int)strlen(_strUTF8), NULL, NULL);
+	MultiByteToWideChar(CP_UTF8, 0, _strUTF8, (int)strlen(_strUTF8), _strUnicode, nLen);
 }
 
 //{
